step3 v1.013 intervals2
