+++ frontend/src/components/plots.js
@@ -1,5 +1,11 @@
+import {CustomHttp} from "../services/custom-http";
+
 export class Plots {
     constructor() {
+        const date = new Date(), y = date.getFullYear(), m = date.getMonth();
+        const today = new Date(y, m, date.getDate() + 1).toISOString().slice(0, 10);
+
+        this.getOperations(today, today);
 
         // chart JS
         const ctxIncome = document.getElementById('myChartIncome');
@@ -58,5 +64,32 @@ export class Plots {
         });
 
         //
+    }
+
+    async getOperations(dateFrom, dateTo) {
+        // 2024-08-23
+        const result = await CustomHttp.request('/operations?period=interval&dateFrom=' + dateFrom + '&dateTo=' + dateTo);
+        // console.log(result.response);
+        if (result) {
+            if (result.redirect) {
+                return this.openNewRoute(result.redirect);
+            }
+            if (result.error || !result.response || result.response.error) {
+                return alert('Возникла ошибка при запросе операций. Обратитесь в поддержку');
+            }
+        }
+
+        console.log(result.response);
+        const items = result.response;
+        // return result.response;
+        this.getIncomes(items);
+    }
+
+    async getIncomes(items) {
+        const incomes = items.filter(item => item.type === 'income');
+        console.log(incomes);
+
+        // incomes.forEach(income => )
+
     }
 }
