+++ frontend/src/templates/pages/dashboard.html
@@ -1,12 +1,12 @@
 <section class="main-section">
     <h1 class="main-title">Главная</h1>
     <div class="flow-period">
-        <a class="flow-period-common flow-period-today flow-period-active">Сегодня</a>
-        <a class="flow-period-common flow-period-week">Неделя</a>
-        <a class="flow-period-common flow-period-month">Месяц</a>
-        <a class="flow-period-common flow-period-year">Год</a>
-        <a class="flow-period-common flow-period-year">Все</a>
-        <a class="flow-period-common flow-period-interval">Интервал</a>
+        <a id="today-main" class="flow-period-common flow-period-today flow-period-active">Сегодня</a>
+        <a id="today-week" class="flow-period-common flow-period-week">Неделя</a>
+        <a id="today-month" class="flow-period-common flow-period-month">Месяц</a>
+        <a id="year-main" class="flow-period-common flow-period-year">Год</a>
+        <a id="all-period-main" class="flow-period-common flow-period-year">Все</a>
+        <a id="interval-main" class="flow-period-common flow-period-interval">Интервал</a>
         <div class="flow-period-select">
             <div class="flow-period-select-since-text">c</div>
             <div class="flow-period-select-since-data"><a>Дата</a></div>
